+++ assets/javascripts/kitten/components/buttons/button-image-with-text-and-badge.js
@@ -1,11 +1,15 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import classNames from 'classnames'
+import COLORS from '../../constants/colors-config'
 
 export class ButtonImageWithTextAndBadge extends React.Component {
   renderNotificationBadge() {
     return (
-      <span className="k-ButtonWithBadge__badge k-Badge">
+      <span
+        className="k-ButtonWithBadge__badge k-Badge"
+        style={{ backgroundColor: this.props.badgeColor }}
+      >
         {this.props.notifications}
       </span>
     )
@@ -69,12 +73,14 @@ ButtonImageWithTextAndBadge.propTypes = {
   widthImg: PropTypes.number,
   heightImg: PropTypes.number,
   altImg: PropTypes.string,
-  notifications: PropTypes.number,
+  notifications: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
+  badgeColor: PropTypes.string,
 }
 
 ButtonImageWithTextAndBadge.defaultProps = {
   children: 'Toggle button',
   onClick: () => {},
+  badgeColor: COLORS.primary1,
 }
 
 // DEPRECATED: do not use default export.